# docs/api/consolidated_api_docs.html/auth-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Auth Service",
    "description": "Authentication and authorization service for Cryptobot",
    "version": "1.0.0"
  },
  "paths": {
...
    "/auth/api-keys": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "List Api Keys",
        "description": "List API keys for current user",
        "operationId": "list_api_keys_auth_api_keys_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/APIKeyOut"
                  },
                  "type": "array",
                  "title": "Response List Api Keys Auth Api Keys Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create Api Key",
        "description": "Create a new API key for exchanges",
        "operationId": "create_api_key_auth_api_keys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/auth__models__user__APIKeyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/auth__schemas__token__APIKeyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api-keys/api-keys": {
      "get": {
        "tags": [
          "api-keys",
          "api-keys"
        ],
        "summary": "List Api Keys",
        "description": "List all API keys for the current user\n\nReturns a list of all API keys owned by the current user,\nincluding active, rotating, and expired keys.",
        "operationId": "list_api_keys_api_keys_api_keys_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "api-keys",
          "api-keys"
        ],
        "summary": "Create Api Key",
        "description": "Create a new API key\n\nCreates a new API key for the specified exchange with the given description.\nThe key will expire after the specified number of days (default: 90).",
        "operationId": "create_api_key_api_keys_api_keys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/auth__schemas__api_key__APIKeyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/auth__schemas__api_key__APIKeyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIKeyListResponse": {
        "properties": {
          "keys": {
            "items": {
              "$ref": "#/components/schemas/auth__schemas__api_key__APIKeyResponse"
            },
            "type": "array",
            "title": "Keys",
            "description": "List of API keys"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of keys"
          },
          "active": {
            "type": "integer",
            "title": "Active",
            "description": "Number of active keys"
          },
          "rotating": {
            "type": "integer",
            "title": "Rotating",
            "description": "Number of keys in rotation"
          },
          "expiring_soon": {
            "type": "integer",
            "title": "Expiring Soon",
            "description": "Number of keys expiring within 7 days"
          }
        },
        "type": "object",
        "required": [
          "keys",
          "total",
          "active",
          "rotating",
          "expiring_soon"
        ],
        "title": "APIKeyListResponse",
        "description": "Response containing a list of API keys"
      },
      "APIKeyOut": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description"
          },
          "exchange": {
            "type": "string",
            "title": "Exchange"
          },
          "is_test": {
            "type": "boolean",
            "title": "Is Test",
            "default": false
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "key": {
            "type": "string",
            "title": "Key"
          }
        },
        "type": "object",
        "required": [
          "description",
          "exchange",
          "id",
          "key"
        ],
        "title": "APIKeyOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KeyStatus": {
        "type": "string",
        "enum": [
          "active",
          "rotating",
          "expired",
          "revoked",
          "compromised"
        ],
        "title": "KeyStatus",
        "description": "API Key status enum"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "auth__models__user__APIKeyCreate": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description"
          },
          "exchange": {
            "type": "string",
            "title": "Exchange"
          },
          "is_test": {
            "type": "boolean",
            "title": "Is Test",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "description",
          "exchange"
        ],
        "title": "APIKeyCreate"
      },
      "auth__schemas__api_key__APIKeyCreate": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Description of the key's purpose"
          },
          "exchange": {
            "type": "string",
            "title": "Exchange",
            "description": "Exchange this key is for"
          },
          "is_test": {
            "type": "boolean",
            "title": "Is Test",
            "description": "Whether this is a test key",
            "default": false
          },
          "expiry_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiry Days",
            "description": "Days until key expires",
            "default": 90
          }
        },
        "type": "object",
        "required": [
          "description",
          "exchange"
        ],
        "title": "APIKeyCreate",
        "description": "Request to create a new API key"
      },
      "auth__schemas__api_key__APIKeyResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique identifier for the key"
          },
          "key": {
            "type": "string",
            "title": "Key",
            "description": "The API key value"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Description of the key's purpose"
          },
          "exchange": {
            "type": "string",
            "title": "Exchange",
            "description": "Exchange this key is for"
          },
          "is_test": {
            "type": "boolean",
            "title": "Is Test",
            "description": "Whether this is a test key"
          },
          "status": {
            "$ref": "#/components/schemas/KeyStatus",
            "description": "Current status of the key"
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "description": "Version number of the key"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "When the key was created"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At",
            "description": "When the key expires"
          },
          "last_used": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used",
            "description": "When the key was last used"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions",
            "description": "Permissions granted to this key"
          },
          "rotated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rotated At",
            "description": "When the key was rotated"
          },
          "grace_period_ends": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grace Period Ends",
            "description": "When the grace period ends"
          },
          "next_key_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Key Id",
            "description": "ID of the next key version"
          },
          "previous_key_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Key Id",
            "description": "ID of the previous key version"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At",
            "description": "When the key was revoked"
          },
          "revocation_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revocation Reason",
            "description": "Reason for revocation"
          },
          "compromised_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compromised At",
            "description": "When the key was marked compromised"
          },
          "compromise_details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compromise Details",
            "description": "Details about the compromise"
          }
        },
        "type": "object",
        "required": [
          "id",
          "key",
          "description",
          "exchange",
          "is_test",
          "status",
          "version",
          "created_at",
          "expires_at",
          "permissions"
        ],
        "title": "APIKeyResponse",
        "description": "Response containing API key details"
      },
      "auth__schemas__token__APIKeyResponse": {
        "properties": {
          "api_key": {
            "type": "string",
            "title": "Api Key"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "api_key",
          "created_at"
        ],
        "title": "APIKeyResponse"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  }
}